#98 Break root_test.cpp up
Fix #98
