+++ test/root_test.cpp
@@ -1,11 +1,7 @@
 /* Copyright (C) 2022 by Camden Mannett.  All rights reserved. */
 
-#include "arg_router/help.hpp"
-#include "arg_router/policy/custom_parser.hpp"
-#include "arg_router/policy/description.hpp"
 #include "arg_router/policy/min_max_value.hpp"
 #include "arg_router/policy/validator.hpp"
-#include "arg_router/policy/value_separator.hpp"
 #include "arg_router/utility/compile_time_string.hpp"
 
 #include "test_helpers.hpp"
@@ -22,335 +18,9 @@ namespace
 {
 using default_validator_type =
     std::decay_t<decltype(policy::validation::default_validator)>;
-
-struct A {
-    explicit A(int v = 0) : value{v} {}
-    int value;
-};
-
-struct B {
-    explicit B(double v = 0) : value{v} {}
-    double value;
-};
 }  // namespace
 
-template <>
-struct arg_router::parser<B> {
-    static B parse(std::string_view token)
-    {
-        return B{parser<double>::parse(token)};
-    }
-};
-
 BOOST_AUTO_TEST_SUITE(root_suite)
-
-BOOST_AUTO_TEST_CASE(is_tree_node_test)
-{
-    static_assert(
-        is_tree_node_v<root_t<flag_t<policy::long_name_t<S_("hello")>,
-                                     policy::router<std::function<void(bool)>>>,
-                              default_validator_type>>,
-        "Tree node test has failed");
-}
-
-BOOST_AUTO_TEST_CASE(validator_type_test)
-{
-    static_assert(
-        std::is_same_v<
-            typename root_t<flag_t<policy::long_name_t<S_("hello")>,
-                                   policy::router<std::function<void(bool)>>>,
-                            default_validator_type>::validator_type,
-            default_validator_type>,
-        "Tree node test has failed");
-}
-
-BOOST_AUTO_TEST_CASE(constructor_validation_test)
-{
-    const auto r = root(policy::validation::default_validator,
-                        flag(policy::long_name<S_("hello")>,
-                             policy::description<S_("This is a hello")>,
-                             policy::short_name<'h'>,
-                             policy::router{[]() {}}),
-                        flag(policy::long_name<S_("goodbye")>,
-                             policy::description<S_("This is a goodbye flag")>,
-                             policy::short_name<'g'>,
-                             policy::router{[]() {}}));
-
-    BOOST_CHECK_EQUAL(std::get<0>(r.children()).long_name(), "hello");
-    BOOST_CHECK_EQUAL(std::get<1>(r.children()).long_name(), "goodbye");
-}
-
-BOOST_AUTO_TEST_CASE(unknown_argument_parse_test)
-{
-    auto router_hit = false;
-    const auto r = root(flag(policy::long_name<S_("hello")>,
-                             policy::description<S_("Hello description")>,
-                             policy::router{[&](bool) { router_hit = true; }}),
-                        policy::validation::default_validator);
-
-    auto args = std::vector{"foo", "--foo"};
-    BOOST_CHECK_EXCEPTION(
-        r.parse(args.size(), const_cast<char**>(args.data())),
-        parse_exception,
-        [](const auto& e) { return e.what() == "Unknown argument: --foo"s; });
-    BOOST_CHECK(!router_hit);
-}
-
-BOOST_AUTO_TEST_CASE(unhandled_parse_test)
-{
-    auto router_hit = false;
-    const auto r = root(flag(policy::long_name<S_("hello")>,
-                             policy::description<S_("Hello description")>,
-                             policy::router{[&](bool) { router_hit = true; }}),
-                        policy::validation::default_validator);
-
-    auto args = std::vector{"foo", "--hello", "--foo"};
-    BOOST_CHECK_EXCEPTION(r.parse(args.size(), const_cast<char**>(args.data())),
-                          parse_exception,
-                          [](const auto& e) {
-                              return e.what() == "Unhandled arguments: --foo"s;
-                          });
-    BOOST_CHECK(!router_hit);
-}
-
-BOOST_AUTO_TEST_CASE(single_flag_parse_test)
-{
-    auto router_hit = false;
-    const auto r = root(flag(policy::long_name<S_("hello")>,
-                             policy::description<S_("Hello description")>,
-                             policy::router{[&](bool) { router_hit = true; }}),
-                        policy::validation::default_validator);
-
-    auto args = std::vector{"foo", "--hello"};
-    r.parse(args.size(), const_cast<char**>(args.data()));
-    BOOST_CHECK(router_hit);
-}
-
-BOOST_AUTO_TEST_CASE(single_arg_parse_test)
-{
-    auto result = std::optional<int>{};
-    const auto r = root(arg<int>(policy::long_name<S_("hello")>,
-                                 policy::description<S_("Hello description")>,
-                                 policy::router{[&](int value) {
-                                     BOOST_CHECK(!result);
-                                     result = value;
-                                 }}),
-                        policy::validation::default_validator);
-
-    auto args = std::vector{"foo", "--hello", "42"};
-    r.parse(args.size(), const_cast<char**>(args.data()));
-    BOOST_REQUIRE(!!result);
-    BOOST_CHECK_EQUAL(*result, 42);
-}
-
-BOOST_AUTO_TEST_CASE(single_arg_separator_parse_test)
-{
-    auto result = std::optional<int>{};
-    const auto r = root(arg<int>(policy::long_name<S_("hello")>,
-                                 policy::description<S_("Hello description")>,
-                                 policy::value_separator<'='>,
-                                 policy::router{[&](int value) {
-                                     BOOST_CHECK(!result);
-                                     result = value;
-                                 }}),
-                        policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected, auto fail_message) {
-        result.reset();
-
-        try {
-            r.parse(args.size(), const_cast<char**>(args.data()));
-            BOOST_CHECK(fail_message.empty());
-            BOOST_REQUIRE(!!result);
-            BOOST_CHECK_EQUAL(*result, expected);
-        } catch (parse_exception& e) {
-            BOOST_CHECK_EQUAL(e.what(), fail_message);
-        }
-    };
-
-    test::data_set(
-        f,
-        {
-            std::tuple{std::vector{"foo", "--hello=42"}, 42, ""s},
-            std::tuple{std::vector{"foo", "--hello", "42"},
-                       0,
-                       "Expected to find value separator '=' in \"--hello\""s},
-            std::tuple{std::vector{"foo", "--hello="},
-                       0,
-                       "Unable to find value after separator: --hello="s},
-        });
-}
-
-BOOST_AUTO_TEST_CASE(single_string_arg_parse_test)
-{
-    auto result = std::optional<std::string_view>{};
-    const auto r =
-        root(arg<std::string_view>(policy::long_name<S_("hello")>,
-                                   policy::description<S_("Hello description")>,
-                                   policy::router{[&](std::string_view value) {
-                                       BOOST_CHECK(!result);
-                                       result = value;
-                                   }}),
-             policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected) {
-        result.reset();
-
-        r.parse(args.size(), const_cast<char**>(args.data()));
-        BOOST_REQUIRE(!!result);
-        BOOST_CHECK_EQUAL(*result, expected);
-    };
-
-    test::data_set(
-        f,
-        {
-            std::tuple{std::vector{"foo", "--hello", "hello"}, "hello"sv},
-            std::tuple{std::vector{"foo", "--hello", "-h"}, "-h"sv},
-            std::tuple{std::vector{"foo", "--hello", "-hello"}, "-hello"sv},
-            std::tuple{std::vector{"foo", "--hello", "--hello"}, "--hello"sv},
-        });
-}
-
-BOOST_AUTO_TEST_CASE(triple_flag_parse_test)
-{
-    auto result = std::array<bool, 3>{};
-    const auto r = root(flag(policy::long_name<S_("flag1")>,
-                             policy::description<S_("First description")>,
-                             policy::router{[&](bool) { result[0] = true; }}),
-                        flag(policy::long_name<S_("flag2")>,
-                             policy::description<S_("Second description")>,
-                             policy::router{[&](bool) { result[1] = true; }}),
-                        flag(policy::short_name<'t'>,
-                             policy::description<S_("Third description")>,
-                             policy::router{[&](bool) { result[2] = true; }}),
-                        policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected) {
-        result.fill(false);
-
-        r.parse(args.size(), const_cast<char**>(args.data()));
-        BOOST_CHECK_EQUAL(result[0], expected[0]);
-        BOOST_CHECK_EQUAL(result[1], expected[1]);
-        BOOST_CHECK_EQUAL(result[2], expected[2]);
-    };
-
-    test::data_set(f,
-                   {
-                       std::tuple{std::vector{"foo", "--flag1"},
-                                  std::array{true, false, false}},
-                       std::tuple{std::vector{"foo", "--flag2"},
-                                  std::array{false, true, false}},
-                       std::tuple{std::vector{"foo", "-t"},
-                                  std::array{false, false, true}},
-                   });
-}
-
-BOOST_AUTO_TEST_CASE(triple_arg_parse_test)
-{
-    auto result = std::tuple<int, double, std::string_view>{};
-    auto hit = std::array<bool, 3>{};
-
-    const auto r =
-        root(arg<int>(policy::long_name<S_("flag1")>,
-                      policy::description<S_("First description")>,
-                      policy::router{[&](auto value) {
-                          std::get<0>(result) = value;
-                          hit[0] = true;
-                      }}),
-             arg<double>(policy::long_name<S_("flag2")>,
-                         policy::description<S_("Second description")>,
-                         policy::router{[&](auto value) {
-                             std::get<1>(result) = value;
-                             hit[1] = true;
-                         }}),
-             arg<std::string_view>(policy::short_name<'t'>,
-                                   policy::description<S_("Third description")>,
-                                   policy::router{[&](auto value) {
-                                       std::get<2>(result) = value;
-                                       hit[2] = true;
-                                   }}),
-             policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected_hit, auto expected_value) {
-        result = decltype(result){};
-        hit.fill(false);
-
-        r.parse(args.size(), const_cast<char**>(args.data()));
-        BOOST_CHECK_EQUAL(hit[0], expected_hit[0]);
-        BOOST_CHECK_EQUAL(hit[1], expected_hit[1]);
-        BOOST_CHECK_EQUAL(hit[2], expected_hit[2]);
-
-        BOOST_CHECK_EQUAL(std::get<0>(result), std::get<0>(expected_value));
-        BOOST_CHECK_EQUAL(std::get<1>(result), std::get<1>(expected_value));
-        BOOST_CHECK_EQUAL(std::get<2>(result), std::get<2>(expected_value));
-    };
-
-    test::data_set(f,
-                   {
-                       std::tuple{std::vector{"foo", "--flag1", "42"},
-                                  std::array{true, false, false},
-                                  std::tuple{42, 0.0, ""sv}},
-                       std::tuple{std::vector{"foo", "--flag2", "3.14"},
-                                  std::array{false, true, false},
-                                  std::tuple{0, 3.14, ""sv}},
-                       std::tuple{std::vector{"foo", "-t", "hello"},
-                                  std::array{false, false, true},
-                                  std::tuple{0, 0.0, "hello"sv}},
-                   });
-}
-
-BOOST_AUTO_TEST_CASE(custom_parser_test)
-{
-    auto result = std::tuple<A, B, B>{};
-    auto parser_hit = false;
-
-    const auto r = root(
-        arg<A>(policy::long_name<S_("arg1")>,
-               policy::description<S_("First description")>,
-               policy::custom_parser<A>{[](auto token) -> A {
-                   return A{parser<int>::parse(token)};
-               }},
-               policy::router{[&](auto arg1) { std::get<0>(result) = arg1; }}),
-        arg<B>(policy::long_name<S_("arg2")>,
-               policy::description<S_("Second description")>,
-               policy::custom_parser<B>{[&](auto token) -> B {
-                   parser_hit = true;
-                   return B{parser<double>::parse(token)};
-               }},
-               policy::router{[&](auto arg2) { std::get<1>(result) = arg2; }}),
-        arg<B>(policy::long_name<S_("arg3")>,
-               policy::description<S_("Third description")>,
-               policy::router{[&](auto arg3) { std::get<2>(result) = arg3; }}),
-        policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected_hit, auto expected_value) {
-        result = decltype(result){A{}, B{}, B{}};
-        parser_hit = false;
-
-        r.parse(args.size(), const_cast<char**>(args.data()));
-        BOOST_CHECK_EQUAL(parser_hit, expected_hit);
-
-        BOOST_CHECK_EQUAL(std::get<0>(result).value,
-                          std::get<0>(expected_value).value);
-        BOOST_CHECK_EQUAL(std::get<1>(result).value,
-                          std::get<1>(expected_value).value);
-        BOOST_CHECK_EQUAL(std::get<2>(result).value,
-                          std::get<2>(expected_value).value);
-    };
-
-    test::data_set(f,
-                   {
-                       std::tuple{std::vector{"foo", "--arg1", "42"},
-                                  false,
-                                  std::tuple{A{42}, B{}, B{}}},
-                       std::tuple{std::vector{"foo", "--arg2", "3.14"},
-                                  true,
-                                  std::tuple{A{}, B{3.14}, B{}}},
-                       std::tuple{std::vector{"foo", "--arg3", "3.3"},
-                                  false,
-                                  std::tuple{A{}, B{}, B{3.3}}},
-                   });
-}
 
 BOOST_AUTO_TEST_CASE(anonymous_mode_single_flag_parse_test)
 {
@@ -530,7 +200,7 @@ BOOST_AUTO_TEST_CASE(multiple_required_arg_parse_test)
                           });
 }
 
-BOOST_AUTO_TEST_CASE(anonymous_triple_single_flag_parse_test)
+BOOST_AUTO_TEST_CASE(anonymous_triple_flag_parse_test)
 {
     auto router_hit = false;
     auto result = std::array<bool, 3>{};
@@ -1015,215 +685,6 @@ BOOST_AUTO_TEST_CASE(alias_arg_parse_test)
         });
 }
 
-BOOST_AUTO_TEST_CASE(single_positional_arg_parse_test)
-{
-    auto router_hit = false;
-    auto result = std::tuple<bool, bool, int, std::vector<std::string_view>>{};
-    const auto r =
-        root(mode(flag(policy::long_name<S_("flag1")>,
-                       policy::short_name<'a'>,
-                       policy::description<S_("First description")>),
-                  flag(policy::short_name<'b'>),
-                  arg<int>(policy::long_name<S_("arg1")>,
-                           policy::default_value(42),
-                           policy::description<S_("Second description")>),
-                  positional_arg<std::vector<std::string_view>>(
-                      policy::display_name<S_("pos_args")>,
-                      policy::description<S_("Third description")>,
-                      policy::required,
-                      policy::min_count<2>),
-                  policy::router{[&](bool flag1,
-                                     bool b,
-                                     int arg1,
-                                     std::vector<std::string_view> pos_args) {
-                      result = decltype(result){flag1, b, arg1, pos_args};
-                      router_hit = true;
-                  }}),
-             policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected, std::string fail_message) {
-        result = decltype(result){};
-        router_hit = false;
-
-        try {
-            r.parse(args.size(), const_cast<char**>(args.data()));
-            BOOST_CHECK(fail_message.empty());
-            BOOST_CHECK(router_hit);
-            BOOST_CHECK_EQUAL(std::get<0>(result), std::get<0>(expected));
-            BOOST_CHECK_EQUAL(std::get<1>(result), std::get<1>(expected));
-            BOOST_CHECK_EQUAL(std::get<2>(result), std::get<2>(expected));
-            BOOST_CHECK(std::get<3>(result) == std::get<3>(expected));
-        } catch (parse_exception& e) {
-            BOOST_CHECK_EQUAL(fail_message, e.what());
-            BOOST_CHECK(!router_hit);
-        }
-    };
-
-    test::data_set(
-        f,
-        {
-            std::tuple{std::vector{"foo", "one", "two"},
-                       std::tuple{false,
-                                  false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"}},
-                       ""},
-            std::tuple{
-                std::vector{"foo", "--one", "two"},
-                std::tuple{false,
-                           false,
-                           42,
-                           std::vector<std::string_view>{"--one", "two"}},
-                ""},
-            std::tuple{
-                std::vector{"foo", "one", "--two"},
-                std::tuple{false,
-                           false,
-                           42,
-                           std::vector<std::string_view>{"one", "--two"}},
-                ""},
-            std::tuple{std::vector{"foo", "--flag1", "one", "two"},
-                       std::tuple{true,
-                                  false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"}},
-                       ""},
-            std::tuple{std::vector{"foo", "-a", "one", "two"},
-                       std::tuple{true,
-                                  false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"}},
-                       ""},
-            std::tuple{std::vector{"foo", "--flag1", "-b", "one", "two"},
-                       std::tuple{true,
-                                  true,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"}},
-                       ""},
-            std::tuple{std::vector{"foo", "--arg1", "14", "one", "two"},
-                       std::tuple{false,
-                                  false,
-                                  14,
-                                  std::vector<std::string_view>{"one", "two"}},
-                       ""},
-            std::tuple{
-                std::vector{"foo", "--arg1", "14", "--flag1", "one", "two"},
-                std::tuple{true,
-                           false,
-                           14,
-                           std::vector<std::string_view>{"one", "two"}},
-                ""},
-            std::tuple{
-                std::vector{"foo", "--arg1", "14", "--flag1", "one", "-two"},
-                std::tuple{true,
-                           false,
-                           14,
-                           std::vector<std::string_view>{"one", "-two"}},
-                ""},
-            std::tuple{
-                std::vector{"foo", "-ab", "--one", "two"},
-                std::tuple{true,
-                           true,
-                           42,
-                           std::vector<std::string_view>{"--one", "two"}},
-                ""},
-            std::tuple{
-                std::vector{"foo", "--flag1", "hello"},
-                std::tuple{true, false, 42, std::vector<std::string_view>{}},
-                "Minimum count not reached: pos_args"},
-            std::tuple{
-                std::vector{"foo", "--flag1", "--arg1", "9", "hello"},
-                std::tuple{true, false, 9, std::vector<std::string_view>{}},
-                "Minimum count not reached: pos_args"},
-            std::tuple{
-                std::vector{"foo", "--flag1"},
-                std::tuple{true, false, 42, std::vector<std::string_view>{}},
-                "Missing required argument: pos_args"},
-        });
-}
-
-BOOST_AUTO_TEST_CASE(two_positional_arg_parse_test)
-{
-    auto router_hit = false;
-    auto result = std::
-        tuple<bool, int, std::vector<std::string_view>, std::vector<double>>{};
-    const auto r = root(
-        mode(flag(policy::long_name<S_("flag1")>,
-                  policy::description<S_("First description")>),
-             arg<int>(policy::long_name<S_("arg1")>,
-                      policy::default_value(42),
-                      policy::description<S_("Second description")>),
-             positional_arg<std::vector<std::string_view>>(
-                 policy::display_name<S_("pos_args1")>,
-                 policy::description<S_("Third description")>,
-                 policy::fixed_count<2>),
-             positional_arg<std::vector<double>>(
-                 policy::display_name<S_("pos_args2")>,
-                 policy::description<S_("Fourth description")>),
-             policy::router{[&](bool flag1,
-                                int arg1,
-                                std::vector<std::string_view> pos_args1,
-                                std::vector<double> pos_args2) {
-                 result = decltype(result){flag1, arg1, pos_args1, pos_args2};
-                 router_hit = true;
-             }}),
-        policy::validation::default_validator);
-
-    auto f = [&](auto args, auto expected, std::string fail_message) {
-        result = decltype(result){};
-        router_hit = false;
-
-        try {
-            r.parse(args.size(), const_cast<char**>(args.data()));
-            BOOST_CHECK(fail_message.empty());
-            BOOST_CHECK(router_hit);
-            BOOST_CHECK_EQUAL(std::get<0>(result), std::get<0>(expected));
-            BOOST_CHECK_EQUAL(std::get<1>(result), std::get<1>(expected));
-            BOOST_CHECK(std::get<2>(result) == std::get<2>(expected));
-            BOOST_CHECK(std::get<3>(result) == std::get<3>(expected));
-        } catch (parse_exception& e) {
-            BOOST_CHECK_EQUAL(fail_message, e.what());
-            BOOST_CHECK(!router_hit);
-        }
-    };
-
-    test::data_set(
-        f,
-        {
-            std::tuple{std::vector{"foo", "one", "two"},
-                       std::tuple{false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"},
-                                  std::vector<double>{}},
-                       ""},
-            std::tuple{std::vector{"foo", "one", "two", "3.14"},
-                       std::tuple{false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"},
-                                  std::vector<double>{3.14}},
-                       ""},
-            std::tuple{std::vector{"foo", "one", "two", "3.14", "443.34"},
-                       std::tuple{false,
-                                  42,
-                                  std::vector<std::string_view>{"one", "two"},
-                                  std::vector<double>{3.14, 443.34}},
-                       ""},
-            std::tuple{std::vector{"foo", "one", "two", "three"},
-                       std::tuple{false,
-                                  42,
-                                  std::vector<std::string_view>{},
-                                  std::vector<double>{}},
-                       "Failed to parse: three"},
-            std::tuple{
-                std::vector{"foo", "one", "--flag1", "two", "--arg1", "5"},
-                std::tuple{false,
-                           42,
-                           std::vector<std::string_view>{},
-                           std::vector<double>{}},
-                "Failed to parse: two"},
-        });
-}
-
 BOOST_AUTO_TEST_CASE(nested_mode_test)
 {
     auto router_hit = std::bitset<6>{};
@@ -1590,284 +1051,5 @@ BOOST_AUTO_TEST_CASE(counting_flag_test)
                        "Dependent argument missing: -a"},
         });
 }
-
-BOOST_AUTO_TEST_CASE(help_test)
-{
-    auto f = [](const auto& root, const auto& expected_result) {
-        const auto result = root.help();
-        BOOST_CHECK_EQUAL(result, expected_result);
-    };
-
-    test::data_set(
-        f,
-        std::tuple{
-            std::tuple{root(flag(policy::long_name<S_("flag1")>,
-                                 policy::short_name<'a'>,
-                                 policy::description<S_("Flag1 description")>,
-                                 policy::router{[](bool) {}}),
-                            flag(policy::long_name<S_("flag2")>,
-                                 policy::router{[](bool) {}}),
-                            flag(policy::short_name<'b'>,
-                                 policy::description<S_("b description")>,
-                                 policy::router{[](bool) {}}),
-                            policy::validation::default_validator),
-                       ""s},
-            std::tuple{
-                root(flag(policy::long_name<S_("flag1")>,
-                          policy::short_name<'a'>,
-                          policy::description<S_("Flag1 description")>,
-                          policy::router{[](bool) {}}),
-                     flag(policy::long_name<S_("flag2")>,
-                          policy::router{[](bool) {}}),
-                     flag(policy::short_name<'b'>,
-                          policy::description<S_("b description")>,
-                          policy::router{[](bool) {}}),
-                     arg<int>(policy::long_name<S_("arg1")>,
-                              policy::value_separator<'='>,
-                              policy::router{[](int) {}}),
-                     help(policy::long_name<S_("help")>,
-                          policy::short_name<'h'>,
-                          policy::description<S_("Help output")>,
-                          policy::program_name<S_("foo")>,
-                          policy::program_version<S_("v3.14")>,
-                          policy::program_intro<S_("My foo is good for you")>),
-                     policy::validation::default_validator),
-                R"(foo v3.14
-
-My foo is good for you
-
-    --flag1,-a        Flag1 description
-    --flag2
-    -b                b description
-    --arg1=<Value>
-    --help,-h         Help output
-)"s},
-            std::tuple{
-                root(help(policy::long_name<S_("help")>,
-                          policy::short_name<'h'>,
-                          policy::description<S_("Help output")>,
-                          policy::program_name<S_("foo")>,
-                          policy::program_version<S_("v3.14")>,
-                          policy::program_intro<S_("My foo is good for you")>),
-                     mode(flag(policy::long_name<S_("flag1")>,
-                               policy::short_name<'a'>,
-                               policy::description<S_("Flag1 description")>),
-                          flag(policy::long_name<S_("flag2")>),
-                          arg<int>(policy::long_name<S_("arg1")>,
-                                   policy::value_separator<'='>,
-                                   policy::description<S_("Arg1 description")>),
-                          flag(policy::short_name<'b'>,
-                               policy::description<S_("b description")>),
-                          policy::router{[](bool, bool, bool) {}}),
-                     policy::validation::default_validator),
-                R"(foo v3.14
-
-My foo is good for you
-
-    --help,-h             Help output
-        --flag1,-a        Flag1 description
-        --flag2
-        --arg1=<Value>    Arg1 description
-        -b                b description
-)"s}});
-}
-
-BOOST_AUTO_TEST_SUITE(death_suite)
-
-BOOST_AUTO_TEST_CASE(must_have_validator_policy_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/policy/validator.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-int main() {
-    arg_router::root_t<
-        arg_router::flag_t<
-            arg_router::policy::short_name_t<
-                arg_router::traits::integral_constant<'a'>>,
-            arg_router::policy::long_name_t<S_("test")>,
-            arg_router::policy::router<std::less<>>>>();
-    return 0;
-}
-    )",
-        "Root must have a validator policy, use "
-        "policy::validation::default_validator unless you have created a "
-        "custom one");
-}
-
-BOOST_AUTO_TEST_CASE(must_not_have_value_separator_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/policy/validator.hpp"
-#include "arg_router/policy/value_separator.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-using default_validator_type =
-    std::decay_t<decltype(policy::validation::default_validator)>;
-
-int main() {
-    root_t<
-        flag_t<
-            policy::short_name_t<traits::integral_constant<'a'>>,
-            policy::long_name_t<S_("test")>,
-            policy::router<std::less<>>>,
-        policy::value_separator_t<traits::integral_constant<'='>>,
-        default_validator_type>();
-    return 0;
-}
-    )",
-        "Root must not have a value separator policy");
-}
-
-BOOST_AUTO_TEST_CASE(must_have_at_least_one_child_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/policy/validator.hpp"
-
-using namespace arg_router;
-
-using default_validator_type =
-    std::decay_t<decltype(policy::validation::default_validator)>;
-
-int main() {
-    arg_router::root_t<default_validator_type>();
-    return 0;
-}
-    )",
-        "Root must have at least one child");
-}
-
-BOOST_AUTO_TEST_CASE(single_child_must_have_router_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/policy/validator.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-using default_validator_type =
-    std::decay_t<decltype(policy::validation::default_validator)>;
-
-int main() {
-    arg_router::root_t<default_validator_type,
-                       flag_t<policy::long_name_t<S_("f1")>>>();
-    return 0;
-}
-    )",
-        "All root children must have routers, unless they have no value");
-}
-
-BOOST_AUTO_TEST_CASE(pre_parse_phase_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/flag.hpp"
-#include "arg_router/policy/alias.hpp"
-#include "arg_router/policy/long_name.hpp"
-#include "arg_router/root.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-int main() {
-    const auto m = root(policy::alias(policy::long_name<S_("other-mode")>),
-                        flag(policy::long_name<S_("hello")>));
-    return 0;
-}
-    )",
-        "Root does not support policies with any parsing phases");
-}
-
-BOOST_AUTO_TEST_CASE(parse_phase_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/flag.hpp"
-#include "arg_router/policy/custom_parser.hpp"
-#include "arg_router/policy/long_name.hpp"
-#include "arg_router/root.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-int main() {
-    const auto m = root(policy::custom_parser<int>{[](auto) { return false; }},
-                        flag(policy::long_name<S_("hello")>));
-    return 0;
-}
-    )",
-        "Root does not support policies with any parsing phases");
-}
-
-BOOST_AUTO_TEST_CASE(validation_phase_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/flag.hpp"
-#include "arg_router/policy/long_name.hpp"
-#include "arg_router/policy/min_max_value.hpp"
-#include "arg_router/root.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-int main() {
-    const auto m = root(policy::min_max_value{1, 3},
-                        flag(policy::long_name<S_("hello")>));
-    return 0;
-}
-    )",
-        "Root does not support policies with any parsing phases");
-}
-
-BOOST_AUTO_TEST_CASE(routing_phase_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/flag.hpp"
-#include "arg_router/policy/long_name.hpp"
-#include "arg_router/policy/router.hpp"
-#include "arg_router/root.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-int main() {
-    const auto m = root(policy::router{[](std::string_view) { return true; }},
-                        flag(policy::long_name<S_("hello")>));
-    return 0;
-}
-    )",
-        "Root does not support policies with any parsing phases");
-}
-
-BOOST_AUTO_TEST_CASE(missing_phase_test)
-{
-    test::death_test_compile(
-        R"(
-#include "arg_router/flag.hpp"
-#include "arg_router/policy/long_name.hpp"
-#include "arg_router/policy/required.hpp"
-#include "arg_router/root.hpp"
-#include "arg_router/utility/compile_time_string.hpp"
-
-using namespace arg_router;
-
-int main() {
-    const auto m = root(policy::required,
-                        flag(policy::long_name<S_("hello")>));
-    return 0;
-}
-    )",
-        "Root does not support policies with any parsing phases");
-}
-
-BOOST_AUTO_TEST_SUITE_END()
 
 BOOST_AUTO_TEST_SUITE_END()
